+++ battery_orchestrator/run.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/with-contenv bashio
 cd /app
-exec python3 main.py
+exec python3 core_app.py